requirements.txt
cffi==1.15.1
click==8.1.7
cryptography==43.0.1
Flask==2.2.5
Flask-Cors==5.0.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.5
pycparser==2.21
PyJWT==2.8.0
python-dotenv==0.21.1
typing_extensions==4.7.1
websockets==11.0.3
Werkzeug==2.2.3
zipp==3.15.0
